M3-113 Feat : myPageController 에서 걸음 관련된 부분 구현
